Move utilities to common
