+++ src/common/util/strings.ts
